refactor(qwen): consolidate external test file into inline tests module
